[ch60624] add unit tests for extrapolation
