+++ tests/python/requirements.txt
@@ -1,0 +1,5 @@
+pytest== 4.6.11
+scipy==1.2.1
+pandas==0.23.4
+numpy==1.16.4
+future==0.18.2
