+++ tests/python/unit/dku_timeseries/test_extrema_extraction.py
