+++ tests/python/unit/dku_timeseries/test_windowing.py
